wirouting: create exception file
