+++ pyroute2/wirouting/__init__.py
@@ -1,5 +1,13 @@
+from pyroute2.wirouting.exception import (
+    InterfaceDoesNotExist,
+    InterfaceExists,
+    NotPhyDevice,
+)
 from pyroute2.wirouting.wiroute import WiRoute
 
-__all__= [
+__all__ = [
     "WiRoute",
+    "InterfaceDoesNotExist",
+    "InterfaceExists",
+    "NotPhyDevice",
 ]
